refactor(rules): single shipped rules corpus with closed severity enum
Replace the two-file proposed/validated corpus topology with ONE shipped
rules.yaml per engine and rework the loader to match:

- git-mv config/engine_invariants/ -> config/engine_rules/ (history kept)
- loader binds to exactly src/llenergymeasure/engines/<e>/rules.yaml; the
  validated-YAML overlay merge and its observed-* vocabulary are removed
- severity is a CLOSED {error, dormant} enum; anything else in a corpus
  file is a load-time RuleCorpusError (warn / dormant_silent / llm_proposed
  machinery dies)
- every rule carries a required provenance block (source, verified,
  citation, engine_version, date); provenance is metadata only and never
  influences runtime behaviour
- corpus seeded by translating main's arbitrated rules for the current
  pins (transformers 41, vllm 10, tensorrt 3); match predicates, severity,
  message templates and normalisation targets unchanged
- consumer contract preserved: try_match / render_message / severity
  dispatch / normalised_fields (via a deprecated expected_outcome shim for
  library_resolution and report_gaps until their B-phase rework)
- consumers updated to the new import path only (models, study,
  api/report_gaps, infra/version_handshake); report_gaps now owns its
  EmissionChannel runtime taxonomy
- engine-pipeline workflow: drop the three dead invariants.*.yaml dorny
  filter lines so corpus deletions cannot fire the old mine+validate cells
- tests reworked to the new contract; corpus-schema validation tests cover
  the closed severity enum and required provenance
