+++ .github/workflows/engine-pipeline.yml
@@ -145,7 +145,6 @@ jobs:
               - 'scripts/validate_invariants.py'
               - 'scripts/_invariant_validation_common.py'
               - 'scripts/refresh_invariant_rules.sh'
-              - 'src/llenergymeasure/engines/transformers/invariants.*.yaml'
               - '.github/workflows/engine-pipeline.yml'
               - '.github/workflows/_engine-invariants-cell.yml'
               - '.github/actions/**'
@@ -158,7 +157,6 @@ jobs:
               - 'scripts/validate_invariants.py'
               - 'scripts/_invariant_validation_common.py'
               - 'scripts/refresh_invariant_rules.sh'
-              - 'src/llenergymeasure/engines/vllm/invariants.*.yaml'
               - '.github/workflows/engine-pipeline.yml'
               - '.github/workflows/_engine-invariants-cell.yml'
               - '.github/actions/**'
@@ -171,7 +169,6 @@ jobs:
               - 'scripts/validate_invariants.py'
               - 'scripts/_invariant_validation_common.py'
               - 'scripts/refresh_invariant_rules.sh'
-              - 'src/llenergymeasure/engines/tensorrt/invariants.*.yaml'
               - '.github/workflows/engine-pipeline.yml'
               - '.github/workflows/_engine-invariants-cell.yml'
               - '.github/actions/**'
