+++ pyproject.toml
@@ -55,7 +55,7 @@ llem = "llenergymeasure.cli:app"
 [tool.hatch.build.targets.wheel]
 packages = ["src/llenergymeasure"]
 # Hatchling includes non-Python files in package dirs by default, so the
-# Z-engines runtime data (src/llenergymeasure/engines/<engine>/{invariants.*.yaml,
+# Z-engines runtime data (src/llenergymeasure/engines/<engine>/{rules.yaml,
 # schema.discovered.json}) ships automatically in the wheel - no force-include
 # needed (and adding one creates duplicates).
 #
