--- src/llenergymeasure/engines/tensorrt/invariants.validated.yaml
@@ -1,45 +1,0 @@
-schema_version: 1.0.0
-engine: tensorrt
-engine_version: 0.21.0
-image_ref: llenergymeasure:tensorrt
-base_image_ref: llenergymeasure:tensorrt
-validated_at: '2026-05-15T12:49:19+02:00'
-validation_commit: 92da87c2f83a30cd4528edfa934f3c1bf5165daf
-cases:
-- id: tensorrt_raises_max_ngram_size_le_0_positive_values
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValidationError
-    message: "1 validation error for LookaheadDecodingConfig\nmax_ngram_size\n  Value\
-      \ error, Value must be positive, got 0 [type=value_error, input_value=0, input_type=int]\n\
-      \    For further information visit https://errors.pydantic.dev/2.11/v/value_error"
-  positive_confirmed: true
-  negative_confirmed: true
-- id: tensorrt_raises_max_verification_set_size_le_0_positive_values
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValidationError
-    message: "1 validation error for LookaheadDecodingConfig\nmax_verification_set_size\n\
-      \  Value error, Value must be positive, got 0 [type=value_error, input_value=0,\
-      \ input_type=int]\n    For further information visit https://errors.pydantic.dev/2.11/v/value_error"
-  positive_confirmed: true
-  negative_confirmed: true
-- id: tensorrt_raises_max_window_size_le_0_positive_values
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValidationError
-    message: "1 validation error for LookaheadDecodingConfig\nmax_window_size\n  Value\
-      \ error, Value must be positive, got 0 [type=value_error, input_value=0, input_type=int]\n\
-      \    For further information visit https://errors.pydantic.dev/2.11/v/value_error"
-  positive_confirmed: true
-  negative_confirmed: true
-divergences: []
