--- src/llenergymeasure/engines/vllm/invariants.validated.yaml
@@ -1,109 +1,0 @@
-schema_version: 1.0.0
-engine: vllm
-engine_version: 0.7.3
-image_ref: llenergymeasure:vllm
-base_image_ref: llenergymeasure:vllm
-validated_at: '2026-05-15T12:49:19+02:00'
-validation_commit: 92da87c2f83a30cd4528edfa934f3c1bf5165daf
-cases:
-- id: vllm_samplingparams_dormant_seed_eq_neg1
-  outcome: dormant_silent
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations:
-    seed:
-      declared: -1
-      observed: null
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_samplingparams_raises_best_of_lt_ref_n
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: best_of must be greater than or equal to n, got n=2 and best_of=1.
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_samplingparams_raises_min_tokens_gt_ref_max_tokens
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: min_tokens must be less than or equal to max_tokens=1, got 2.
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_samplingparams_raises_min_tokens_lt_0
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: min_tokens must be greater than or equal to 0, got -1.
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_samplingparams_raises_n_lt_1
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: n must be at least 1, got 0.
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_samplingparams_raises_n_not_type_int
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: n must be an int, but is of type <class 'str'>
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_samplingparams_raises_temperature_lt_0p0
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: temperature must be non-negative, got -1.0.
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_schedulerconfig_raises_max_num_partial_prefills_lt_1
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: max_num_partial_prefills (0) must be greater than or equal to 1.
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_schedulerconfig_raises_num_lookahead_slots_lt_0
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: num_lookahead_slots (-1) must be greater than or equal to 0.
-  positive_confirmed: true
-  negative_confirmed: true
-- id: vllm_schedulerconfig_raises_num_scheduler_steps_lt_1
-  outcome: error
-  emission_channel: none
-  observed_messages: []
-  observed_silent_normalisations: {}
-  observed_exception:
-    type: ValueError
-    message: num_scheduler_steps (0) must be greater than or equal to 1.
-  positive_confirmed: true
-  negative_confirmed: true
-divergences: []
