--- tests/unit/config/engine_invariants/test_corpus_invariants.py
@@ -1,229 +1,0 @@
-"""Invariants enforced on the seeded validation-invariants corpus.
-
-The validation CI pipeline (a separate follow-up PR) runs a richer equivalent
-of these checks - empirically verifying each invariant's ``kwargs_positive``
-actually fires in the real library and ``kwargs_negative`` doesn't. These
-schema-level invariants are the offline backstop.
-
-Most enum coverage is redundant with the loader's own ``UnknownEnumValueError``
-family (severity / outcome / emission_channel / added_by are all validated at
-corpus parse time). These tests re-iterate for defence-in-depth and as a
-human-readable catalogue of what the corpus must look like.
-"""
-
-from __future__ import annotations
-
-from pathlib import Path
-
-import pytest
-
-from llenergymeasure.config.engine_invariants import (
-    VALID_ADDED_BY,
-    VALID_EMISSION_CHANNEL,
-    VALID_OUTCOME,
-    VALID_SEVERITY,
-    EngineInvariantsLoader,
-)
-
-
-@pytest.fixture(scope="module")
-def transformers_corpus():
-    loader = EngineInvariantsLoader()
-    return loader.load_invariants("transformers")
-
-
-def test_corpus_covers_required_invariants(transformers_corpus) -> None:
-    """Coverage-by-invariant: every required field surface has at least one invariant.
-
-    Pins SEMANTIC coverage, not specific invariant IDs. Extractor refinements
-    that rename invariants don't break this test; extractor regressions that
-    drop coverage of a real invariant do. See
-    ``.product/designs/config-deduplication-dormancy/runtime-config-validation.md``
-    decision-log entry of 2026-04-25 (corpus-as-measurement principle).
-    """
-    invariants = transformers_corpus.invariants
-
-    def covers_field(field_path: str) -> bool:
-        return any(field_path in invariant.match_fields for invariant in invariants)
-
-    # Single-field invariants - at least one invariant must touch each path.
-    # Scope reflects the regenerated canonical corpus produced by
-    # ``scripts/engine_producers/build_corpus.py`` with the validation gate.
-    # Adding a path here means: an invariant for that field must survive validation
-    # against the pinned engine library (see ``engine_version``
-    # in the corpus envelope). If a path drops out, investigate WHY (real
-    # extractor regression, real library change, or validation-harness gap)
-    # before weakening this list.
-    required_fields = (
-        # Greedy dormancy: do_sample=False / num_beams=1 strip these.
-        "transformers.sampling.temperature",
-        "transformers.sampling.top_p",
-        "transformers.sampling.top_k",
-        "transformers.sampling.min_p",
-        "transformers.sampling.typical_p",
-        "transformers.sampling.epsilon_cutoff",
-        "transformers.sampling.eta_cutoff",
-        # Single-beam dormancy.
-        "transformers.sampling.early_stopping",
-        "transformers.sampling.length_penalty",
-        # Note: num_beam_groups + diversity_penalty validations were softened
-        # in transformers 4.57.x (error → dormant_announced or no_op). The
-        # validation-CI gate quarantines the corpus invariants that previously claimed
-        # error severity. Coverage loss tracked separately; do NOT re-add
-        # without first confirming the library re-introduced enforcement.
-        # No-return-dict dormancy.
-        "transformers.sampling.output_scores",
-        "transformers.sampling.output_attentions",
-        "transformers.sampling.output_hidden_states",
-        # GenerationConfig.validate() error invariants.
-        "transformers.sampling.max_new_tokens",
-        "transformers.sampling.cache_implementation",
-        "transformers.sampling.num_return_sequences",
-        "transformers.sampling.pad_token_id",
-        "transformers.sampling.compile_config",
-        # PR #387 cross-field invariant gates.
-        "transformers.sampling.num_beams",
-        # Watermarking + BNB type-check paths landed in PR 5 (validation-
-        # validation gate). The BNB invariants use the field paths the real
-        # ExperimentConfig schema exposes (``transformers.load_in_4bit``
-        # etc.) rather than the old ``transformers.quant.<field>`` paths
-        # which never resolved at runtime.
-        "transformers.sampling.watermarking_config",
-        "transformers.load_in_4bit",
-        "transformers.load_in_8bit",
-        "transformers.llm_int8_threshold",
-        "transformers.llm_int8_skip_modules",
-        "transformers.llm_int8_enable_fp32_cpu_offload",
-        "transformers.llm_int8_has_fp16_weight",
-        # Note: transformers.bnb_4bit_compute_dtype - validation-CI quarantined
-        # the type-check invariant under 4.57.3. Coverage loss tracked in a
-        # follow-up alongside num_beam_groups / diversity_penalty.
-        "transformers.bnb_4bit_quant_type",
-        "transformers.bnb_4bit_use_double_quant",
-    )
-    missing = [path for path in required_fields if not covers_field(path)]
-    assert not missing, (
-        f"corpus is missing invariants for {len(missing)} required invariants: {missing}"
-    )
-
-    # Cross-field invariants - at least one invariant must AND-combine the listed
-    # fields. Catches regressions where the extractor lost the cross-field
-    # predicate machinery. PR 5 added the (num_beam_groups,
-    # diversity_penalty) pair (the AST miner's beam-search divisibility
-    # invariant) and the (num_beams, num_return_sequences) pair (the
-    # @field_ref-tightened greedy-rejects predicate).
-    cross_field_pairs = (
-        # Note: (num_beams, num_beam_groups) and (num_beam_groups,
-        # diversity_penalty) cross-field invariants were quarantined in 4.57.x
-        # along with the single-field dormancy invariants; see required_fields
-        # comment above.
-        ("transformers.sampling.num_beams", "transformers.sampling.num_return_sequences"),
-    )
-    missing_pairs = [
-        pair
-        for pair in cross_field_pairs
-        if not any(all(p in invariant.match_fields for p in pair) for invariant in invariants)
-    ]
-    assert not missing_pairs, (
-        f"corpus missing cross-field invariants for {len(missing_pairs)} invariants: {missing_pairs}"
-    )
-
-    # Corpus-as-measurement: the regenerated corpus is machine-extracted by
-    # construction, so no invariant should carry ``added_by: manual_seed``. Any
-    # manual entry indicates a hand-edit of the canonical YAML that bypasses
-    # the build_corpus.py + validation gate and would silently drift
-    # on the next library bump. This PR's regeneration drops the legacy
-    # hand-curated BNB type-check entries; the AST miner now emits them
-    # under ``added_by: static_miner``.
-    manual = [invariant.id for invariant in invariants if invariant.added_by == "manual_seed"]
-    assert not manual, (
-        f"corpus contains {len(manual)} hand-seeded invariants; corpus must be "
-        f"machine-extracted (run scripts/engine_producers/build_corpus.py): {manual}"
-    )
-
-
-def test_corpus_schema_version_is_current(transformers_corpus) -> None:
-    assert transformers_corpus.schema_version.startswith("1.")
-
-
-def test_corpus_ids_unique(transformers_corpus) -> None:
-    ids = [invariant.id for invariant in transformers_corpus.invariants]
-    assert len(ids) == len(set(ids))
-
-
-def test_corpus_match_fields_non_empty(transformers_corpus) -> None:
-    for invariant in transformers_corpus.invariants:
-        assert invariant.match_fields, f"Invariant {invariant.id} has empty match.fields"
-
-
-def test_corpus_kwargs_positive_and_negative_populated(transformers_corpus) -> None:
-    for invariant in transformers_corpus.invariants:
-        assert invariant.kwargs_positive, f"Invariant {invariant.id} has empty kwargs_positive"
-        assert invariant.kwargs_negative, f"Invariant {invariant.id} has empty kwargs_negative"
-
-
-def test_corpus_severity_values_are_valid(transformers_corpus) -> None:
-    for invariant in transformers_corpus.invariants:
-        assert invariant.severity in VALID_SEVERITY, invariant.id
-
-
-def test_corpus_outcome_values_are_valid(transformers_corpus) -> None:
-    for invariant in transformers_corpus.invariants:
-        outcome = invariant.expected_outcome.get("outcome")
-        assert outcome in VALID_OUTCOME, f"{invariant.id}: bad outcome {outcome!r}"
-
-
-def test_corpus_emission_channels_valid(transformers_corpus) -> None:
-    for invariant in transformers_corpus.invariants:
-        channel = invariant.expected_outcome.get("emission_channel")
-        assert channel in VALID_EMISSION_CHANNEL, f"{invariant.id}: {channel!r}"
-
-
-def test_corpus_severity_outcome_consistency(transformers_corpus) -> None:
-    for invariant in transformers_corpus.invariants:
-        outcome = invariant.expected_outcome["outcome"]
-        if invariant.severity == "error":
-            assert outcome == "error", f"{invariant.id}: severity=error but outcome={outcome}"
-        elif invariant.severity == "warn":
-            assert outcome == "warn", f"{invariant.id}: severity=warn but outcome={outcome}"
-        elif invariant.severity == "dormant":
-            assert outcome in {"dormant_silent", "dormant_announced"}, (
-                f"{invariant.id}: severity=dormant but outcome={outcome}"
-            )
-
-
-def test_corpus_dormant_silent_has_normalised_fields(transformers_corpus) -> None:
-    for invariant in transformers_corpus.invariants:
-        if invariant.expected_outcome.get("outcome") == "dormant_silent":
-            normalised = invariant.expected_outcome.get("normalised_fields")
-            assert normalised, f"{invariant.id}: dormant_silent but no normalised_fields"
-
-
-def test_corpus_added_by_values_valid(transformers_corpus) -> None:
-    # Defers to the loader's single source of truth (VALID_ADDED_BY) - the
-    # loader's UnknownAddedByError would have rejected the invariant at parse
-    # time if the value were outside the enum. Kept here for human-readable
-    # catalogue + defence-in-depth.
-    for invariant in transformers_corpus.invariants:
-        assert invariant.added_by in VALID_ADDED_BY, (
-            f"{invariant.id}: added_by={invariant.added_by!r}"
-        )
-
-
-def test_corpus_file_is_valid_yaml() -> None:
-    # Sanity: the on-disk file is parseable YAML (redundant with loader, but
-    # guards against accidental corruption by direct edits).
-    import yaml
-
-    path = (
-        Path(__file__).resolve().parents[4]
-        / "src"
-        / "llenergymeasure"
-        / "engines"
-        / "transformers"
-        / "invariants.proposed.yaml"
-    )
-    assert path.exists()
-    doc = yaml.safe_load(path.read_text())
-    assert isinstance(doc, dict)
-    assert "invariants" in doc
